+++ src/main/java/me/drex/worldmanager/gui/create/SelectSeed.java
@@ -3,6 +3,7 @@ package me.drex.worldmanager.gui.create;
 import eu.pb4.sgui.api.elements.GuiElementBuilder;
 import eu.pb4.sgui.api.gui.AnvilInputGui;
 import eu.pb4.sgui.api.gui.SimpleGui;
+import me.drex.worldmanager.gui.util.GuiElements;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.item.Items;
 
@@ -15,10 +16,7 @@ public class SelectSeed extends AnvilInputGui {
     public SelectSeed(ServerPlayer player, SimpleGui previousGui, Consumer<Long> consumer) {
         super(player, false);
         setTitle(localized("worldmanager.gui.select_seed.title"));
-        setSlot(1, new GuiElementBuilder(Items.OAK_DOOR)
-            .setName(localized("worldmanager.gui.generic.back.name"))
-            .setCallback(previousGui::open)
-        );
+        setSlot(1, GuiElements.back(previousGui));
         setSlot(2, new GuiElementBuilder(Items.GREEN_STAINED_GLASS_PANE)
             .setName(localized("worldmanager.gui.select_seed.confirm.name"))
             .setCallback(() -> {